refactor: move core logic into internal/program package
