+++ internal/program/exec_test.go
@@ -1,4 +1,4 @@
-package main
+package program
 
 import (
 	"fmt"
