+++ internal/program/types.go
@@ -1,0 +1,13 @@
+package program
+
+type Device struct {
+	Name      string
+	Connected bool
+}
+
+func getSymbol(status bool) string {
+	if status {
+		return "󰂱"
+	}
+	return "󰂲"
+}
